requirements.txt
antiorm==1.2.1
asgiref==3.7.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
db==0.1.1
db-sqlite3==0.0.1
distlib==0.3.7
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
platformdirs==3.11.0
protobuf==4.21.12
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.6
Werkzeug==3.0.1
